% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicpv-package.R
\docType{package}
\name{dynamicpv-package}
\alias{dynamicpv}
\alias{dynamicpv-package}
\title{dynamicpv: Evaluates Present Values and Cost-Effectiveness with Dynamic Pricing and Uptake}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Evaluates present values and cost-effectiveness with dynamic pricing and uptake.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://MSDLLCpapers.github.io/dynamicpv/}
  \item \url{https://github.com/MSDLLCpapers/dynamicpv}
  \item Report bugs at \url{https://github.com/MSDLLCpapers/dynamicpv/issues}
}

}
\author{
\strong{Maintainer}: Dominic Muston \email{dominic.muston@msd.com} (\href{https://orcid.org/0000-0003-4876-7940}{ORCID})

Other contributors:
\itemize{
  \item John Blischak (\href{https://orcid.org/0000-0003-2634-9879}{ORCID}) [contributor]
  \item Merck & Co., Inc., Rahway, NJ, USA and its affiliates [copyright holder, funder]
}

}
\keyword{internal}
